+++ CHANGELOG.md
@@ -19,6 +19,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Fix behaviour of `--preserve-structure` when using internal or external axiom selectors for [`remove`] or [`filter`] [#816]
 - Fix value rendering for entities in [`report`] [#874]
 
+### Changed
+- Do not allow malformed IRIs to be returned by `IOHelper` [#882]
+
 ## [1.8.1] - 2021-01-27
 
 ### Fixed
@@ -257,6 +260,7 @@ First official release of ROBOT!
 [`template`]: http://robot.obolibrary.org/template
 [`validate`]: http://robot.obolibrary.org/validate
 
+[#882]: https://github.com/ontodev/robot/pull/882
 [#874]: https://github.com/ontodev/robot/pull/874
 [#872]: https://github.com/ontodev/robot/pull/872
 [#858]: https://github.com/ontodev/robot/pull/858